+++ qucs/contrib/qucs-release.sh
@@ -196,4 +196,8 @@ tar -zcvhf qucs-$RELEASE.tar.gz qucs-$RELEASE
 rm -rf qucs-$RELEASE
 tar -zxvf qucs-$RELEASE.tar.gz #make the symbolic links actual files
 
+echo Saving copy to ${HOME}/Downloads/qucs-$RELEASE.tar.gz
+cp qucs-$RELEASE.tar.gz ${HOME}/Downloads/
+
+